code/requirements.txt
flask==2.1.0
google-api-python-client==2.42.0
google-cloud-logging==3.0.0
google-cloud-core==2.2.3
oauth2client==4.1.3
google-cloud-compute==1.1.0
google-cloud-container==2.10.7
google-cloud-gke-hub==1.4.3
kubernetes==23.3.0